+++ regions/shapes/tests/test_rectangle.py
@@ -135,12 +135,17 @@ class TestRectanglePixelRegion(BaseTestPixelRegion):
         # For now this will only work with unrotated rectangles. Once
         # this works with rotated rectangles, the following exception
         # check can be removed as well as the ``angle=0 * u.deg`` in the
-        # call to copy() below.
-        with pytest.raises(NotImplementedError,
-                           match=('Cannot create matplotlib selector for rotated rectangle.')):
-            self.reg.as_mpl_selector(ax)
+        # copy() below - should (hopefully) be implemented with mpl 3.6.
+        if MPL_VERSION < 36:
+            with pytest.raises(NotImplementedError,
+                               match=('Cannot create matplotlib selector for rotated rectangle.')):
+                self.reg.as_mpl_selector(ax)
 
-        region = self.reg.copy(angle=0 * u.deg)
+            angle = 0 * u.deg
+        else:
+            angle = self.reg.angle
+
+        region = self.reg.copy(angle=angle)
 
         selector = region.as_mpl_selector(ax, callback=update_mask, sync=sync)
 
@@ -164,7 +169,7 @@ class TestRectanglePixelRegion(BaseTestPixelRegion):
             assert_allclose(region.center.y, 4)
             assert_allclose(region.width, 4)
             assert_allclose(region.height, 3)
-            assert_quantity_allclose(region.angle, 0 * u.deg)
+            assert_quantity_allclose(region.angle, angle)
 
             assert_equal(mask, 0)